requirements.txt
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
ChatterBot==1.0.8
chatterbot-corpus==1.2.0
click==7.1.2
cymem==2.0.3
dnspython==2.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Bcrypt==0.7.1
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
MarkupSafe==1.1.1
mathparse==0.1.2
murmurhash==1.0.2
nltk==3.5
numpy==1.19.2
plac==1.1.3
preshed==3.0.2
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
PyJWT==1.7.1
PyYAML==3.13
regex==2020.9.27
requests==2.24.0
six==1.15.0
spacy==2.3.2
SQLAlchemy==1.3.19
srsly==1.0.2
thinc==7.4.1
tqdm==4.50.0
urllib3==1.25.10
wasabi==0.8.0
Werkzeug==1.0.1
zipp==3.2.0


python -m spacy download en